Split out Units class into separate header
- run format
